Replace snapshot-based IPA detection with dictionary comparison
detect-ipa-changes.ts now compares words.ts against espeak-ng (en-us)
American English dictionary instead of a stored snapshot. Removed the
snapshot file and verify-ipa.ts (superseded).

Usage:
  bun run scripts/detect-ipa-changes.ts           # all differences
  bun run scripts/detect-ipa-changes.ts --major    # consonant diffs only
  bun run scripts/detect-ipa-changes.ts --delete   # delete audio for major mismatches
